refactor: 댓글 컨트롤러 단수, 복수로 컨트롤러 분리 및 테스트 코드 수정
